Using print
The Basics of Writing a Program

We all interact with many kinds of devices, websites, and apps that are made from "programs." As you might imagine, apps such as Instagram or Google Maps are quite complex and have taken teams of professional programmers many months or years to create. You are starting on the journey that can lead to creating programs that are just polished and interactive, but we will begin with much simpler programs.

The essence of all programs, no matter how complex or simple, is that they provide a step-by-step list of instructions that the computer will follow. The computer, while powerful, is also very simplistic—it only does what you tell it. In addition, it does exactly what you tell it to do, nothing more and nothing less, and if you leave out a very small thing, or make even a tiny mistake, the computer is very unforgiving.

Starting with this activity, you are going to practice writing programs—the step-by-step instructions for the computer—and you are going to start with simple ones. You will start by displaying words on the screen and have the user type responses in a text window. While not nearly as glamorous as the professional apps you see around you, the fundamentals are the same. These simple programs will help you develop skills you need to create more powerful and more exciting applications as you continue to learn.