Input and Output
In this lesson, I  learned build on the basic input/output ideas from your programs in the last lesson, and learn to store more information and display it using better formatting. While not all programs use the console input and output approach that we have taken thus far (where people type the answers and see the results in text), almost all programs make use of many string variables and combine and format them in one form or another.
As shown in these videos, some helpful string functions available in Python are:

Code	Result
words = "the cat IN THE hat"

the cat IN THE hat

words.capitalize()

The cat in the hat

words.title()

The Cat In The Hat

words.upper()

THE CAT IN THE HAT

words.lower()

the cat in the hat

words.count("t")

3

words.lower().count("t")

4