Create 5_project_clever_stories.py
Overview
Mad Libs are a type of funny story, where a person is asked for words without knowing their context. The words are then placed into a story in a pre-determined format, often resulting in funny statements.

For example, consider prompts for:

Plural noun

Verb

Adjective

Noun

And a story, such as:

When it comes to [plural-noun], you would never want to [verb], especially if you encountered a [adjective] [noun].

A person, may respond to the prompts with the following:

Plural noun: ducks

Verb: jump

Adjective: cold

Noun: taco

Then the story would read:

When it comes to ducks, you would never want to jump, especially if you encountered a cold taco.